64 bit malloc

+++ gatb-core/src/gatb/system/api/IMemory.hpp
@@ -66,7 +66,7 @@ public:
 
     /************************************************************/
     /** Alias for the size of memory blocks handled by the interface.  */
-    typedef u_int32_t BlockSize_t;
+    typedef u_int64_t BlockSize_t;
 
     /** Alias for the size of memory handled by the interface.  */
     typedef u_int64_t TotalSize_t;
